+++ webextension/options.js
@@ -75,8 +75,8 @@ function restoreOptions() {
     document.getElementById('save').textContent = chrome.i18n.getMessage("save");
     document.getElementById('ignoreQuotedLinesDesc').innerHTML = chrome.i18n.getMessage("ignoreQuotedLines");
     document.getElementById('havePremiumAccountDesc').innerHTML = chrome.i18n.getMessage("havePremiumAccountDesc", "https://languagetoolplus.com");
-    document.getElementById('username').innerHTML = chrome.i18n.getMessage("username");
-    document.getElementById('password').innerHTML = chrome.i18n.getMessage("password");
+    document.getElementById('usernameDesc').innerHTML = chrome.i18n.getMessage("username");
+    document.getElementById('passwordDesc').innerHTML = chrome.i18n.getMessage("password");
     document.getElementById('motherTongueDesc').textContent = chrome.i18n.getMessage("motherTongueDesc");
     document.getElementById('motherTongueExpl').textContent = chrome.i18n.getMessage("motherTongueExpl");
     document.getElementById('variant-en-desc').textContent = chrome.i18n.getMessage("variantEnDesc");
@@ -90,8 +90,8 @@ function restoreOptions() {
         apiServerUrl: defaultServerUrl,
         ignoreQuotedLines: true,
         havePremiumAccount: false,
-        username:  "",
-        password:  "",
+        username: "",
+        password: "",
         motherTongue: "",
         enVariant: "en-US",
         deVariant: "de-DE",
@@ -147,10 +147,14 @@ function setPremium(enabled) {
         document.getElementById('password').disabled = false;
         document.getElementById('username').focus();
         document.getElementById('apiServerUrl').disabled = true;
+        document.getElementById('ltPlusAccess').style.display = "block";
+        document.getElementById('ltServer').style.display = "none";
     } else {
         document.getElementById('username').disabled = true;
         document.getElementById('password').disabled = true;
         document.getElementById('apiServerUrl').disabled = false;
+        document.getElementById('ltPlusAccess').style.display = "none";
+        document.getElementById('ltServer').style.display = "block";
     }
 }