+++ webextension/CHANGES.md
@@ -1,5 +1,15 @@
 # LanguageTool WebExtension Change Log
 
+### 1.0.26 (2017-xx-xx)
+  * Added support for using languagetoolplus.com - activate
+    this in the options if you have a premium account there
+    and want the add-on to use it
+  * Added Dutch translation
+
+### 1.0.25 (2017-10-02)
+  * Fixed a bug that made the orientation setting (landscape/portrait) disappear
+    from Chrome's printing dialog
+
 ### 1.0.24 (2017-09-29)
   * Internal preparation for upcoming features
   * Tracking fixes